Update patch set 1
Patch Set 1:

Good point. In this series, I only implemented Bazel build for gitblit.jar. But it can be done much more, like gitblit.war, etc. I havn't looked into it as we don't use it in gerrit. But theoretically, folks, that uses gitblit as standalone product would benefit from maintained Moxie build in our fork.

But I agree, that if we stabilize the way we use gitblit as Git viewer only and Bazel build for gitblit.jar is stable, we may consider to remove Moxie build. I would like to do it later and not in the beginning of Bazel migration.

One reason is QS. By running Moxie, we can easily investigate the content if artifacts created in Bazel and in Moxie build and compare it against each other.

Patch-set: 1
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>